@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
 
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>craftsman</ToolCommandName>
@@ -15,8 +15,9 @@
     <PackageTags>webapi c-sharp dotnet-core vertical-slice-architecture asp-net-core docker net-6 ddd domain-driven-design web-dev cqrs mediatr masstransit event-driven-architecture</PackageTags>
     <RepositoryUrl>https://github.com/pdevito3/craftsman</RepositoryUrl>
     <RepositoryType>git</RepositoryType>
-    <FileVersion>0.13.0-pre2</FileVersion>
-    <Version>0.13.0-pre2</Version>
+    <AssemblyVersion>0.13.0</AssemblyVersion>
+    <FileVersion>0.13.0</FileVersion>
+    <Version>0.13.0</Version>
     <PackageLicenseFile>LICENSE.txt</PackageLicenseFile>
   </PropertyGroup>
 
@@ -27,7 +28,8 @@
     <PackageReference Include="FluentAssertions" Version="5.10.3" />
     <PackageReference Include="FluentValidation" Version="10.3.0" />
     <PackageReference Include="FluentValidation.AspNetCore" Version="10.3.0" />
-    <PackageReference Include="LibGit2Sharp" Version="0.26.2" />
+    <PackageReference Include="Humanizer.Core" Version="2.13.14" />
+    <PackageReference Include="LibGit2Sharp" Version="0.27.0-preview-0175" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="RestSharp" Version="106.12.0" />
     <PackageReference Include="Spectre.Console" Version="0.41.0" />
