@@ -12,21 +12,21 @@
 
     public static class EntityBuilder
     {
-        public static void CreateEntity(string srcDirectory, Entity entity, string projectBaseName, IFileSystem fileSystem)
+        public static void CreateEntity(string solutionDirectory, string srcDirectory, Entity entity, string projectBaseName, IFileSystem fileSystem)
         {
             var classPath = ClassPathHelper.EntityClassPath(srcDirectory, $"{entity.Name}.cs", entity.Plural, projectBaseName);
-            var fileText = GetEntityFileText(classPath.ClassNamespace, srcDirectory, entity, projectBaseName);
-            Utilities.CreateFile(classPath, fileText, fileSystem);
-        }
-        
-        public static void CreateBaseEntity(string srcDirectory, string projectBaseName, IFileSystem fileSystem)
-        {
-            var classPath = ClassPathHelper.EntityClassPath(srcDirectory, $"BaseEntity.cs", "", projectBaseName);
-            var fileText = GetBaseEntityFileText(classPath.ClassNamespace);
+            var fileText = GetEntityFileText(classPath.ClassNamespace, solutionDirectory, srcDirectory, entity, projectBaseName);
             Utilities.CreateFile(classPath, fileText, fileSystem);
         }
 
-        public static string GetEntityFileText(string classNamespace, string srcDirectory, Entity entity, string projectBaseName)
+        public static void CreateBaseEntity(string srcDirectory, string projectBaseName, bool useSoftDelete, IFileSystem fileSystem)
+        {
+            var classPath = ClassPathHelper.EntityClassPath(srcDirectory, $"BaseEntity.cs", "", projectBaseName);
+            var fileText = GetBaseEntityFileText(classPath.ClassNamespace, useSoftDelete);
+            Utilities.CreateFile(classPath, fileText, fileSystem);
+        }
+
+        public static string GetEntityFileText(string classNamespace, string solutionDirectory, string srcDirectory, Entity entity, string projectBaseName)
         {
             var creationDtoName = Utilities.GetDtoName(entity.Name, Dto.Creation);
             var creationValidatorName = Utilities.ValidatorNameGenerator(entity.Name, Validator.Creation);
@@ -36,19 +36,19 @@
             var propString = EntityPropBuilder(entity.Properties);
             var usingSieve = entity.Properties.Where(e => e.CanFilter || e.CanSort).ToList().Count > 0 ? @$"{Environment.NewLine}using Sieve.Attributes;" : "";
             var tableAnnotation = EntityAnnotationBuilder(entity);
-            
+
             var foreignEntityUsings = "";
             var foreignProps = entity.Properties.Where(e => e.IsForeignKey).ToList();
             foreach (var entityProperty in foreignProps)
             {
                 var classPath = ClassPathHelper.EntityClassPath(srcDirectory, $"", entityProperty.ForeignEntityPlural, projectBaseName);
-                
+
                 foreignEntityUsings += $@"
 using {classPath.ClassNamespace};";
             }
 
             var profileClassPath = ClassPathHelper.ProfileClassPath(srcDirectory, $"", entity.Plural, projectBaseName);
-            var dtoClassPath = ClassPathHelper.DtoClassPath(srcDirectory, $"", entity.Name, projectBaseName);
+            var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, $"", entity.Name, projectBaseName);
             var validatorClassPath = ClassPathHelper.ValidationClassPath(srcDirectory, $"", entity.Plural, projectBaseName);
 
             return @$"namespace {classNamespace};
@@ -58,7 +58,7 @@ using {profileClassPath.ClassNamespace};
 using {validatorClassPath.ClassNamespace};
 using AutoMapper;
 using FluentValidation;
-using Newtonsoft.Json;
+using System.Text.Json.Serialization;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Runtime.Serialization;{usingSieve}{foreignEntityUsings}
@@ -93,8 +93,22 @@ public class {entity.Name} : BaseEntity
 }}";
         }
 
-        public static string GetBaseEntityFileText(string classNamespace)
+        public static string GetBaseEntityFileText(string classNamespace, bool useSoftDelete)
         {
+            var isDeletedProp = useSoftDelete
+                ? $@"
+    public bool IsDeleted {{ get; private set; }}"
+                : "";
+
+            var isDeletedMethod = useSoftDelete
+                ? $@"
+    
+    public void UpdateIsDeleted(bool isDeleted)
+    {{
+        IsDeleted = isDeleted;
+    }}"
+                : "";
+
             return @$"namespace {classNamespace};
 
 using System.ComponentModel.DataAnnotations;
@@ -107,7 +121,7 @@ public abstract class BaseEntity
     public DateTime CreatedOn {{ get; private set; }}
     public string? CreatedBy {{ get; private set; }}
     public DateTime? LastModifiedOn {{ get; private set; }}
-    public string? LastModifiedBy {{ get; private set; }}
+    public string? LastModifiedBy {{ get; private set; }}{isDeletedProp}
 
     public void UpdateCreationProperties(DateTime createdOn, string? createdBy)
     {{
@@ -119,7 +133,7 @@ public abstract class BaseEntity
     {{
         LastModifiedOn = lastModifiedOn;
         LastModifiedBy = lastModifiedBy;
-    }}
+    }}{isDeletedMethod}
 }}";
         }
 
@@ -127,7 +141,7 @@ public abstract class BaseEntity
         {
             if (string.IsNullOrEmpty(entity.TableName))
                 return null;
-            
+
             // must have table name to have a schema :-(
             var tableName = entity.TableName;
             return entity.Schema != null ? @$"[Table(""{tableName}"", Schema=""{entity.Schema}"")]" : @$"[Table(""{tableName}"")]";
@@ -144,7 +158,10 @@ public abstract class BaseEntity
                 var newLine = (property.IsForeignKey && !property.IsMany)
                     ? Environment.NewLine
                     : $"{Environment.NewLine}{Environment.NewLine}";
-                propString += $@"    public {property.Type} {property.Name} {{ get; private set; }}{defaultValue}{newLine}";
+
+                if(property.IsPrimativeType || property.IsMany)
+                    propString += $@"    public {property.Type} {property.Name} {{ get; private set; }}{defaultValue}{newLine}";
+
                 propString += GetForeignProp(property);
             }
 
@@ -156,11 +173,14 @@ public abstract class BaseEntity
             var attributeString = "";
             if (entityProperty.IsRequired)
                 attributeString += @$"    [Required]{Environment.NewLine}";
-            if (entityProperty.IsForeignKey && !entityProperty.IsMany)
+            if (entityProperty.IsForeignKey
+                && !entityProperty.IsMany
+                && entityProperty.IsPrimativeType
+            )
                 attributeString += @$"    [JsonIgnore]
     [IgnoreDataMember]
     [ForeignKey(""{entityProperty.ForeignEntityName}"")]{Environment.NewLine}";
-            if(entityProperty.IsMany)
+            if(entityProperty.IsMany || !entityProperty.IsPrimativeType)
                 attributeString += $@"    [JsonIgnore]
     [IgnoreDataMember]{Environment.NewLine}";
             if (entityProperty.CanFilter || entityProperty.CanSort)
