@@ -1,16 +1,14 @@
-﻿namespace Craftsman.Builders
+﻿namespace Craftsman.Builders.Auth
 {
-    using Craftsman.Helpers;
-    using Craftsman.Models;
     using System;
-    using System.Collections.Generic;
     using System.IO;
+    using Helpers;
 
-    public class SwaggerServiceRegistrationModifier
+    public class PermissionsModifier
     {
-        public static void AddPolicies(string srcDirectory, List<Policy> policies, string projectBaseName)
+        public static void AddPermission(string srcDirectory, string permission, string projectBaseName)
         {
-            var classPath = ClassPathHelper.WebApiServiceExtensionsClassPath(srcDirectory, $"{Utilities.GetSwaggerServiceExtensionName()}.cs", projectBaseName);
+            var classPath = ClassPathHelper.PolicyDomainClassPath(srcDirectory, $"Permissions.cs", projectBaseName);
 
             if (!Directory.Exists(classPath.ClassDirectory))
                 Directory.CreateDirectory(classPath.ClassDirectory);
@@ -18,9 +16,9 @@
             if (!File.Exists(classPath.FullClassPath))
                 throw new FileNotFoundException($"The `{classPath.FullClassPath}` file could not be found.");
 
-            var policiesString = "";
-            var nonExistantPolicies = Utilities.GetPoliciesThatDoNotExist(policies, classPath.FullClassPath);
-            policiesString += $@"{Environment.NewLine}{Utilities.GetSwaggerPolicies(nonExistantPolicies)}";
+            var fileText = File.ReadAllText(classPath.FullClassPath);
+            if (fileText.Contains($"const string {permission}"))
+                return;
 
             var tempPath = $"{classPath.FullClassPath}temp";
             using (var input = File.OpenText(classPath.FullClassPath))
@@ -28,18 +26,12 @@
                 using (var output = new StreamWriter(tempPath))
                 {
                     string line;
-                    bool updateNextLine = false;
                     while (null != (line = input.ReadLine()))
                     {
                         var newText = $"{line}";
-                        if (line.Contains($"Scopes ="))
+                        if (line.Contains($"Permissions marker"))
                         {
-                            updateNextLine = true;
-                        }
-                        else if (updateNextLine)
-                        {
-                            newText += policiesString;
-                            updateNextLine = false;
+                            newText += @$"{Environment.NewLine}    public const string {permission} = ""{permission}"";";
                         }
 
                         output.WriteLine(newText);
@@ -53,4 +45,3 @@
         }
     }
 }
-
