@@ -9,13 +9,13 @@
     {
         public static void CreateBasePaginationParameters(string solutionDirectory, string projectBaseName, IFileSystem fileSystem)
         {
-            var classPath = ClassPathHelper.SharedDtoClassPath(solutionDirectory, $"BasePaginationParameters.cs", projectBaseName);
+            var classPath = ClassPathHelper.SharedDtoClassPath(solutionDirectory, $"BasePaginationParameters.cs");
 
             if (!fileSystem.Directory.Exists(classPath.ClassDirectory))
                 fileSystem.Directory.CreateDirectory(classPath.ClassDirectory);
 
             if (fileSystem.File.Exists(classPath.FullClassPath))
-                throw new FileAlreadyExistsException(classPath.FullClassPath);
+                return; // ***change from normal!**** if it exists, don't do anything instead of throwing an error
 
             using (var fs = fileSystem.File.Create(classPath.FullClassPath))
             {
@@ -27,23 +27,25 @@
 
         public static string GetBasePaginationParametersText(string classNamespace)
         {
-            return @$"namespace {classNamespace};
-public abstract class BasePaginationParameters
+            return @$"namespace {classNamespace}
 {{
-    internal virtual int MaxPageSize {{ get; }} = 20;
-    internal virtual int DefaultPageSize {{ get; set; }} = 10;
-
-    public virtual int PageNumber {{ get; set; }} = 1;
-
-    public int PageSize
+    public abstract class BasePaginationParameters
     {{
-        get
+        internal virtual int MaxPageSize {{ get; }} = 20;
+        internal virtual int DefaultPageSize {{ get; set; }} = 10;
+
+        public virtual int PageNumber {{ get; set; }} = 1;
+
+        public int PageSize
         {{
-            return DefaultPageSize;
-        }}
-        set
-        {{
-            DefaultPageSize = value > MaxPageSize ? MaxPageSize : value;
+            get
+            {{
+                return DefaultPageSize;
+            }}
+            set
+            {{
+                DefaultPageSize = value > MaxPageSize ? MaxPageSize : value;
+            }}
         }}
     }}
 }}";
