@@ -12,26 +12,23 @@
     {
         public static string GetReadParameterDtoText(string solutionDirectory, string classNamespace, Entity entity, Dto dto, string projectBaseName)
         {
-            var sharedDtoClassPath = ClassPathHelper.SharedDtoClassPath(solutionDirectory, "", projectBaseName);
+            var sharedDtoClassPath = ClassPathHelper.SharedDtoClassPath(solutionDirectory, "");
 
-            return @$"namespace {classNamespace};
-
-using {sharedDtoClassPath.ClassNamespace};
-
-public class {Utilities.GetDtoName(entity.Name, dto)} : BasePaginationParameters
+            return @$"namespace {classNamespace}
 {{
-    public string Filters {{ get; set; }}
-    public string SortOrder {{ get; set; }}
+    using {sharedDtoClassPath.ClassNamespace};
+
+    public class {Utilities.GetDtoName(entity.Name, dto)} : BasePaginationParameters
+    {{
+        public string Filters {{ get; set; }}
+        public string SortOrder {{ get; set; }}
+    }}
 }}";
         }
 
         public static string GetDtoText(IClassPath dtoClassPath, Entity entity, Dto dto)
         {
-            var propString = dto is Dto.Read ? $@"    public Guid Id {{ get; set; }}
-    public DateTime CreatedOn {{ get; set; }}
-    public string? CreatedBy {{ get; set; }}
-    public DateTime? LastModifiedOn {{ get; set; }}
-    public string? LastModifiedBy {{ get; set; }}{Environment.NewLine}" : "";
+            var propString = dto is Dto.Read ? $@"    public Guid Id {{ get; set; }}{Environment.NewLine}" : "";
             propString += DtoPropBuilder(entity.Properties, dto);
             if (dto is Dto.Update or Dto.Creation)
                 propString = "";
@@ -42,15 +39,15 @@ public class {Utilities.GetDtoName(entity.Name, dto)} : BasePaginationParameters
             if (dto is Dto.Creation or Dto.Update)
                 inheritanceString = $": {Utilities.GetDtoName(entity.Name, Dto.Manipulation)}";
 
-            return @$"namespace {dtoClassPath.ClassNamespace};
-
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-public {abstractString}class {Utilities.GetDtoName(entity.Name, dto)} {inheritanceString}
+            return @$"namespace {dtoClassPath.ClassNamespace}
 {{
-{propString}
+    using System.Collections.Generic;
+    using System;
+
+    public {abstractString}class {Utilities.GetDtoName(entity.Name, dto)} {inheritanceString}
+    {{
+    {propString}
+    }}
 }}";
         }
         
@@ -63,12 +60,14 @@ public {abstractString}class {Utilities.GetDtoName(entity.Name, dto)} {inheritan
                     continue;
                 if (props[eachProp].IsForeignKey && props[eachProp].IsMany)
                     continue;
+                if (!props[eachProp].IsPrimativeType)
+                    continue;
                 var guidDefault = dto == Dto.Creation && props[eachProp].Type.IsGuidPropertyType()
                     ? " = Guid.NewGuid();"
                     : "";
 
                 string newLine = eachProp == props.Count - 1 ? "" : Environment.NewLine;
-                propString += $@"   public {props[eachProp].Type} {props[eachProp].Name} {{ get; set; }}{guidDefault}{newLine}";
+                propString += $@"        public {props[eachProp].Type} {props[eachProp].Name} {{ get; set; }}{guidDefault}{newLine}";
             }
 
             return propString;
