@@ -7,7 +7,7 @@
 
     public class CreateRecordEndpointBuilder
     {
-        public static string GetEndpointTextForCreateRecord(Entity entity, bool addSwaggerComments,List<Policy> policies)
+        public static string GetEndpointTextForCreateRecord(Entity entity, bool addSwaggerComments, Feature feature)
         {
             var lowercaseEntityVariable = entity.Name.LowercaseFirstLetter();
             var entityName = entity.Name;
@@ -16,7 +16,7 @@
             var primaryKeyProp = Entity.PrimaryKeyProperty;
             var addRecordCommandMethodName = Utilities.CommandAddName(entityName);
             var singleResponse = $@"{readDto}";
-            var addRecordAuthorizations = EndpointSwaggerCommentBuilders.BuildAuthorizations(policies);
+            var addRecordAuthorizations = feature.IsProtected ? EndpointSwaggerCommentBuilders.BuildAuthorizations(feature.PermissionName) : "";
             var creationPropName = $"{lowercaseEntityVariable}ForCreation";
 
             return @$"{EndpointSwaggerCommentBuilders.GetSwaggerComments_CreateRecord(entity, addSwaggerComments, singleResponse, addRecordAuthorizations.Length > 0)}{addRecordAuthorizations}
@@ -34,7 +34,7 @@
     }}";
         }
         
-        public static string GetEndpointTextForCreateList(Entity entity, bool addSwaggerComments,List<Policy> policies, Feature feature)
+        public static string GetEndpointTextForCreateList(Entity entity, bool addSwaggerComments, Feature feature)
         {
             var lowercaseEntityVariable = entity.Name.LowercaseFirstLetter();
             var entityName = entity.Name;
@@ -42,7 +42,7 @@
             var creationDto = $"IEnumerable<{Utilities.GetDtoName(entityName, Dto.Creation)}>";
             var primaryKeyProp = Entity.PrimaryKeyProperty;
             var responseObj = $@"IEnumerable<{readDto}>";
-            var addRecordAuthorizations = EndpointSwaggerCommentBuilders.BuildAuthorizations(policies);
+            var addRecordAuthorizations = feature.IsProtected ? EndpointSwaggerCommentBuilders.BuildAuthorizations(feature.PermissionName) : "";
             var batchPropNameLower = feature.BatchPropertyName.LowercaseFirstLetter();
             var creationPropName = $"{lowercaseEntityVariable}ForCreation";
 
