@@ -10,17 +10,17 @@
     using System.IO.Abstractions;
     using System.Linq;
     using System.Text;
+    using Humanizer;
 
     public class SwaggerBuilder
     {
-        public static void AddSwagger(string solutionDirectory, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, IEnumerable<Policy> policies, string projectBaseName, IFileSystem fileSystem)
+        public static void AddSwagger(string solutionDirectory, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, string policyName, string projectBaseName, IFileSystem fileSystem)
         {
-            if (!swaggerConfig.IsSameOrEqualTo(new SwaggerConfig()))
-            {
-                AddSwaggerServiceExtension(solutionDirectory, projectBaseName, swaggerConfig, projectName, addJwtAuthentication, policies, fileSystem);
-                WebApiAppExtensionsBuilder.CreateSwaggerWebApiAppExtension(solutionDirectory, swaggerConfig, addJwtAuthentication, projectBaseName, fileSystem);
-                UpdateWebApiCsProjSwaggerSettings(solutionDirectory, projectBaseName);
-            }
+            if (swaggerConfig.IsSameOrEqualTo(new SwaggerConfig())) return;
+            
+            AddSwaggerServiceExtension(solutionDirectory, projectBaseName, swaggerConfig, projectName, addJwtAuthentication, policyName, fileSystem);
+            WebApiAppExtensionsBuilder.CreateSwaggerWebApiAppExtension(solutionDirectory, swaggerConfig, addJwtAuthentication, projectBaseName, fileSystem);
+            UpdateWebApiCsProjSwaggerSettings(solutionDirectory, projectBaseName);
         }
 
         public static void RegisterSwaggerInStartup(string srcDirectory, string projectBaseName = "")
@@ -61,23 +61,14 @@
             File.Move(tempPath, classPath.FullClassPath);
         }
 
-        public static void AddSwaggerServiceExtension(string srcDirectory, string projectBaseName, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, IEnumerable<Policy> policies, IFileSystem fileSystem)
+        public static void AddSwaggerServiceExtension(string srcDirectory, string projectBaseName, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, string policyName, IFileSystem fileSystem)
         {
             var classPath = ClassPathHelper.WebApiServiceExtensionsClassPath(srcDirectory, $"{Utilities.GetSwaggerServiceExtensionName()}.cs", projectBaseName);
-
-            if (!fileSystem.Directory.Exists(classPath.ClassDirectory))
-                fileSystem.Directory.CreateDirectory(classPath.ClassDirectory);
-
-            if (fileSystem.File.Exists(classPath.FullClassPath))
-                throw new FileAlreadyExistsException(classPath.FullClassPath);
-
-            using var fs = fileSystem.File.Create(classPath.FullClassPath);
-            var data = "";
-            data = GetSwaggerServiceExtensionText(classPath.ClassNamespace, swaggerConfig, projectName, addJwtAuthentication, policies);
-            fs.Write(Encoding.UTF8.GetBytes(data));
+            var fileText = GetSwaggerServiceExtensionText(classPath.ClassNamespace, swaggerConfig, projectName, addJwtAuthentication, policyName);
+            Utilities.CreateFile(classPath,fileText, fileSystem);
         }
 
-        public static string GetSwaggerServiceExtensionText(string classNamespace, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, IEnumerable<Policy> policies)
+        public static string GetSwaggerServiceExtensionText(string classNamespace, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, string policyName)
         {
             return @$"namespace {classNamespace};
 
@@ -95,11 +86,11 @@ using System.Reflection;
 
 public static class SwaggerServiceExtension
 {{
-    {GetSwaggerServiceExtensionText(swaggerConfig, projectName, addJwtAuthentication, policies)}
+    {GetSwaggerServiceExtensionText(swaggerConfig, projectName, addJwtAuthentication, policyName)}
 }}";
         }
 
-        private static string GetSwaggerServiceExtensionText(SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, IEnumerable<Policy> policies)
+        private static string GetSwaggerServiceExtensionText(SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, string policyName)
         {
             var contactUrlLine = IsCleanUri(swaggerConfig.ApiContact.Url)
                 ? $@"
@@ -112,7 +103,6 @@ public static class SwaggerServiceExtension
 
             var licenseText = GetLicenseText(swaggerConfig.LicenseName, licenseUrlLine);
 
-            var policyScopes = Utilities.GetSwaggerPolicies(policies);
             var swaggerAuth = addJwtAuthentication ? $@"
 
             config.AddSecurityDefinition(""oauth2"", new OpenApiSecurityScheme
@@ -125,7 +115,8 @@ public static class SwaggerServiceExtension
                         AuthorizationUrl = new Uri(Environment.GetEnvironmentVariable(""AUTH_AUTHORIZATION_URL"")),
                         TokenUrl = new Uri(Environment.GetEnvironmentVariable(""AUTH_TOKEN_URL"")),
                         Scopes = new Dictionary<string, string>
-                        {{{policyScopes}
+                        {{
+                            {{""{policyName}"", ""{projectName.Humanize()} access""}}
                         }}
                     }}
                 }}
