@@ -11,31 +11,45 @@
 
     public static class FakesBuilder
     {
-        public static void CreateFakes(string solutionDirectory, string projectName, Entity entity, IFileSystem fileSystem)
+        public static void CreateFakes(string solutionDirectory, string testDirectory, string projectBaseName, Entity entity, IFileSystem fileSystem)
         {
             // ****this class path will have an invalid FullClassPath. just need the directory
-            var classPath = ClassPathHelper.TestFakesClassPath(solutionDirectory, $"", entity.Name, projectName);
+            var classPath = ClassPathHelper.TestFakesClassPath(testDirectory, $"", entity.Name, projectBaseName);
 
             if (!Directory.Exists(classPath.ClassDirectory))
                 Directory.CreateDirectory(classPath.ClassDirectory);
 
-            CreateFakerEntityFile(solutionDirectory, entity.Name, entity, projectName, fileSystem);
-            CreateFakerFile(solutionDirectory, Utilities.GetDtoName(entity.Name, Dto.Creation), entity, projectName, fileSystem);
-            CreateFakerFile(solutionDirectory, Utilities.GetDtoName(entity.Name, Dto.Read), entity, projectName, fileSystem);
-            CreateFakerFile(solutionDirectory, Utilities.GetDtoName(entity.Name, Dto.Update), entity, projectName, fileSystem);
+            CreateFakerEntityFile(solutionDirectory, testDirectory, entity.Name, entity, projectBaseName, fileSystem);
+            CreateFakerFile(solutionDirectory, testDirectory, Utilities.GetDtoName(entity.Name, Dto.Creation), entity, projectBaseName, fileSystem);
+            CreateFakerFile(solutionDirectory, testDirectory, Utilities.GetDtoName(entity.Name, Dto.Read), entity, projectBaseName, fileSystem);
+            CreateFakerFile(solutionDirectory, testDirectory, Utilities.GetDtoName(entity.Name, Dto.Update), entity, projectBaseName, fileSystem);
+        }
+        
+        public static void CreateRolePermissionFakes(string solutionDirectory, string testDirectory, string projectBaseName, Entity entity, IFileSystem fileSystem)
+        {
+            var classPath = ClassPathHelper.TestFakesClassPath(testDirectory, $"", entity.Name, projectBaseName);
+
+            if (!Directory.Exists(classPath.ClassDirectory))
+                Directory.CreateDirectory(classPath.ClassDirectory);
+
+            CreateFakerEntityFile(solutionDirectory, testDirectory, entity.Name, entity, projectBaseName, fileSystem);
+            CreateFakerFile(solutionDirectory, testDirectory, Utilities.GetDtoName(entity.Name, Dto.Read), entity, projectBaseName, fileSystem);
+            
+            CreateRolePermissionFakerForCreationOrUpdateFile(solutionDirectory, testDirectory, Utilities.GetDtoName(entity.Name, Dto.Creation), entity, projectBaseName, fileSystem);
+            CreateRolePermissionFakerForCreationOrUpdateFile(solutionDirectory, testDirectory, Utilities.GetDtoName(entity.Name, Dto.Update), entity, projectBaseName, fileSystem);
         }
 
-        private static void CreateFakerFile(string solutionDirectory, string objectToFakeClassName, Entity entity, string projectName, IFileSystem fileSystem)
+        private static void CreateFakerFile(string solutionDirectory, string testDirectory, string objectToFakeClassName, Entity entity, string projectBaseName, IFileSystem fileSystem)
         {
             var fakeFilename = $"Fake{objectToFakeClassName}.cs";
-            var classPath = ClassPathHelper.TestFakesClassPath(solutionDirectory, fakeFilename, entity.Name, projectName);
-            var fileText = GetFakeFileText(classPath.ClassNamespace, objectToFakeClassName, entity, solutionDirectory, projectName);
+            var classPath = ClassPathHelper.TestFakesClassPath(testDirectory, fakeFilename, entity.Name, projectBaseName);
+            var fileText = GetFakeFileText(classPath.ClassNamespace, objectToFakeClassName, entity, solutionDirectory, testDirectory, projectBaseName);
             Utilities.CreateFile(classPath, fileText, fileSystem);
         }
 
-        private static string GetFakeFileText(string classNamespace, string objectToFakeClassName, Entity entity, string solutionDirectory, string projectBaseName)
+        private static string GetFakeFileText(string classNamespace, string objectToFakeClassName, Entity entity, string solutionDirectory, string testDirectory, string projectBaseName)
         {
-            var entitiesClassPath = ClassPathHelper.EntityClassPath(solutionDirectory, "", entity.Plural, projectBaseName);
+            var entitiesClassPath = ClassPathHelper.EntityClassPath(testDirectory, "", entity.Plural, projectBaseName);
             var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, "", entity.Name, projectBaseName);
 
             // this... is super fragile. Should really refactor this
@@ -57,17 +71,17 @@ public class Fake{objectToFakeClassName} : AutoFaker<{objectToFakeClassName}>
 }}";
         }
 
-        private static void CreateFakerEntityFile(string solutionDirectory, string objectToFakeClassName, Entity entity, string projectName, IFileSystem fileSystem)
+        private static void CreateFakerEntityFile(string solutionDirectory, string testDirectory, string objectToFakeClassName, Entity entity, string projectBaseName, IFileSystem fileSystem)
         {
             var fakeFilename = $"Fake{objectToFakeClassName}.cs";
-            var classPath = ClassPathHelper.TestFakesClassPath(solutionDirectory, fakeFilename, entity.Name, projectName);
-            var fileText = GetFakeEntityFileText(classPath.ClassNamespace, objectToFakeClassName, entity, solutionDirectory, projectName);
+            var classPath = ClassPathHelper.TestFakesClassPath(testDirectory, fakeFilename, entity.Name, projectBaseName);
+            var fileText = GetFakeEntityFileText(classPath.ClassNamespace, objectToFakeClassName, entity, solutionDirectory, testDirectory, projectBaseName);
             Utilities.CreateFile(classPath, fileText, fileSystem);
         }
 
-        private static string GetFakeEntityFileText(string classNamespace, string objectToFakeClassName, Entity entity, string solutionDirectory, string projectBaseName)
+        private static string GetFakeEntityFileText(string classNamespace, string objectToFakeClassName, Entity entity, string solutionDirectory, string testDirectory, string projectBaseName)
         {
-            var entitiesClassPath = ClassPathHelper.EntityClassPath(solutionDirectory, "", entity.Plural, projectBaseName);
+            var entitiesClassPath = ClassPathHelper.EntityClassPath(testDirectory, "", entity.Plural, projectBaseName);
             var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, "", entity.Name, projectBaseName);
             var creationDtoName = Utilities.GetDtoName(entity.Name, Dto.Creation);
 
@@ -86,6 +100,33 @@ public class Fake{objectToFakeClassName}
 }}";
         }
         
-        
+
+        private static void CreateRolePermissionFakerForCreationOrUpdateFile(string solutionDirectory, string testDirectory, string objectToFakeClassName, Entity entity, string projectBaseName, IFileSystem fileSystem)
+        {
+            var fakeFilename = $"Fake{objectToFakeClassName}.cs";
+            var classPath = ClassPathHelper.TestFakesClassPath(testDirectory, fakeFilename, entity.Name, projectBaseName);
+            
+            var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, "", entity.Name, projectBaseName);
+            var policyDomainClassPath = ClassPathHelper.PolicyDomainClassPath(testDirectory, "",projectBaseName);
+            var rolesClassPath = ClassPathHelper.SharedKernelDomainClassPath(solutionDirectory, "");
+
+            var fileText= @$"namespace {classPath.ClassNamespace};
+
+using AutoBogus;
+using {policyDomainClassPath.ClassNamespace};
+using {dtoClassPath.ClassNamespace};
+using {rolesClassPath.ClassNamespace};
+
+public class Fake{objectToFakeClassName} : AutoFaker<{objectToFakeClassName}>
+{{
+    public Fake{objectToFakeClassName}()
+    {{
+        RuleFor(rp => rp.Permission, f => f.PickRandom(Permissions.List()));
+        RuleFor(rp => rp.Role, f => f.PickRandom(Roles.List()));
+    }}
+}}";
+
+            Utilities.CreateFile(classPath, fileText, fileSystem);
+        }
     }
 }
