@@ -18,8 +18,6 @@
             if (!File.Exists(classPath.FullClassPath))
                 throw new FileNotFoundException($"The `{classPath.FullClassPath}` file could not be found.");
 
-            var consumerFeatureClassPath = ClassPathHelper.ConsumerFeaturesClassPath(testDirectory, $"", projectBaseName);
-
             var usingsAdded = false;
             var tempPath = $"{classPath.FullClassPath}temp";
             using (var input = File.OpenText(classPath.FullClassPath))
@@ -29,16 +27,18 @@
                 while (null != (line = input.ReadLine()))
                 {
                     var newText = $"{line}";
-                    if (line.Contains($"// MassTransit Setup -- Do Not Delete Comment"))
+                    if (line.Contains($"MassTransit Harness Setup"))
                     {
                         newText += $@"
-        _provider = services.AddMassTransitInMemoryTestHarness(cfg =>
+        services.AddMassTransitInMemoryTestHarness(cfg =>
         {{
             // Consumer Registration -- Do Not Delete Comment
-        }}).BuildServiceProvider();
+        }});";
+                    }
+                    else if (line.Contains($"MassTransit Start Setup"))
+                    {
+                        newText += $@"
         _harness = _provider.GetRequiredService<InMemoryTestHarness>();
-
-        services.AddScoped(_ => Mock.Of<IPublishEndpoint>());
         await _harness.Start();";
                     }
                     else if (line.Contains($"using") && !usingsAdded)
@@ -55,17 +55,68 @@ using MassTransit;";
                     else if (line.Contains($"// MassTransit Methods -- Do Not Delete Comment"))
                     {
                         newText += $@"
-    public static async Task PublishMessage<T>(object message)
-        where T : class
+    /// <summary>
+    /// Publishes a message to the bus, and waits for the specified response.
+    /// </summary>
+    /// <param name=""message"">The message that should be published.</param>
+    /// <typeparam name=""TMessage"">The message that should be published.</typeparam>
+    public static async Task PublishMessage<TMessage>(object message)
+        where TMessage : class
     {{
-        await _harness.Bus.Publish<T>(message);
+        await _harness.Bus.Publish<TMessage>(message);
+    }}
+    
+    /// <summary>
+    /// Confirm if there was a fault when publishing for this harness.
+    /// </summary>
+    /// <typeparam name=""TMessage"">The message that should be published.</typeparam>
+    /// <returns>A boolean of true if there was a fault for a message of the given type when published.</returns>
+    public static async Task<bool> IsFaultyPublished<TMessage>()
+        where TMessage : class
+    {{
+        return await _harness.Published.Any<Fault<TMessage>>();
+    }}
+    
+    /// <summary>
+    /// Confirm that a message has been published for this harness.
+    /// </summary>
+    /// <typeparam name=""TMessage"">The message that should be published.</typeparam>
+    /// <returns>A boolean of true if a message of the given type has been published.</returns>
+    public static async Task<bool> IsPublished<TMessage>()
+        where TMessage : class
+    {{
+        return await _harness.Published.Any<TMessage>();
+    }}
+    
+    /// <summary>
+    /// Confirm that a message has been consumed for this harness.
+    /// </summary>
+    /// <typeparam name=""TMessage"">The message that should be consumed.</typeparam>
+    /// <returns>A boolean of true if a message of the given type has been consumed.</returns>
+    public static async Task<bool> IsConsumed<TMessage>()
+        where TMessage : class
+    {{
+        return await _harness.Consumed.Any<TMessage>();
+    }}
+    
+    /// <summary>
+    /// The desired consumer consumed the message.
+    /// </summary>
+    /// <typeparam name=""TMessage"">The message that should be consumed.</typeparam>
+    /// <typeparam name=""TConsumedBy"">The consumer of the message.</typeparam>
+    /// <returns>A boolean of true if a message of the given type has been consumed by the given consumer.</returns>
+    public static async Task<bool> IsConsumed<TMessage, TConsumedBy>()
+        where TMessage : class
+        where TConsumedBy : class, IConsumer
+    {{
+        var consumerHarness = _provider.GetRequiredService<IConsumerTestHarness<TConsumedBy>>();
+        return await consumerHarness.Consumed.Any<TMessage>();
     }}";
                     }
-                    else if (line.Contains($"private static Checkpoint _checkpoint;"))
+                    else if (line.Contains($"static ServiceProvider _provider;"))
                     {
                         newText += $@"
-    public static InMemoryTestHarness _harness;
-    public static ServiceProvider _provider;";
+    private static InMemoryTestHarness _harness;";
                     }
 
                     output.WriteLine(newText);
@@ -77,7 +128,7 @@ using MassTransit;";
             File.Move(tempPath, classPath.FullClassPath);
         }
 
-        public static void AddMTConsumer(string testDirectory, string consumerName, string projectBaseName, string srcDirectory)
+        public static void AddMTConsumer(string testDirectory, string consumerName, string domainDirectory, string projectBaseName, string srcDirectory)
         {
             var classPath = ClassPathHelper.IntegrationTestProjectRootClassPath(testDirectory, "TestFixture.cs", projectBaseName);
 
@@ -87,7 +138,7 @@ using MassTransit;";
             if (!File.Exists(classPath.FullClassPath))
                 throw new FileNotFoundException($"The `{classPath.FullClassPath}` file could not be found.");
 
-            var consumerClassPath = ClassPathHelper.ConsumerFeaturesClassPath(srcDirectory, $"", projectBaseName);
+            var consumerClassPath = ClassPathHelper.ConsumerFeaturesClassPath(srcDirectory, $"", domainDirectory, projectBaseName);
 
             var tempPath = $"{classPath.FullClassPath}temp";
             var hasUsingForConsumerNamespace = false;
