@@ -62,7 +62,7 @@
                 FileParsingHelper.RunPrimaryKeyGuard(template.Entities);
 
                 // add all files based on the given template config
-                RunEntityBuilders(srcDirectory, testDirectory, template, fileSystem);
+                RunEntityBuilders(solutionDirectory, srcDirectory, testDirectory, template, fileSystem);
 
                 WriteHelpHeader($"{Environment.NewLine}Your entities have been successfully added. Keep up the good work!");
             }
@@ -94,25 +94,24 @@
             }
         }
 
-        private static void RunEntityBuilders(string srcDirectory, string testDirectory, AddEntityTemplate template, IFileSystem fileSystem)
+        private static void RunEntityBuilders(string solutionDirectory, string srcDirectory, string testDirectory, AddEntityTemplate template, IFileSystem fileSystem)
         {
+            var projectBaseName = template.SolutionName;
+            var useSoftDelete = Utilities.ProjectUsesSoftDelete(srcDirectory, projectBaseName);
+            
             //entities
-            EntityScaffolding.ScaffoldEntities(srcDirectory,
+            EntityScaffolding.ScaffoldEntities(solutionDirectory,
+                srcDirectory,
                 testDirectory,
-                template.SolutionName,
+                projectBaseName,
                 template.Entities,
                 template.DbContextName,
                 template.AddSwaggerComments,
+                useSoftDelete,
                 fileSystem);
-            
-            //TODO move inside scaffoldentites?
-            foreach (var feature in template.Entities.SelectMany(entity => entity.Features))
-            {
-                InfrastructureServiceRegistrationModifier.AddPolicies(srcDirectory, feature.Policies , template.SolutionName);
-                SwaggerServiceRegistrationModifier.AddPolicies(srcDirectory, feature.Policies , template.SolutionName);
-            }
-            SeederModifier.AddSeeders(srcDirectory, template.Entities, template.DbContextName, template.SolutionName);
-            DbContextModifier.AddDbSet(srcDirectory, template.Entities, template.DbContextName, template.SolutionName);
+
+            SeederModifier.AddSeeders(srcDirectory, template.Entities, template.DbContextName, projectBaseName);
+            DbContextModifier.AddDbSet(srcDirectory, template.Entities, template.DbContextName, projectBaseName);
         }
 
         private static AddEntityTemplate GetDbContext(string srcDirectory, AddEntityTemplate template, string projectBaseName)
