@@ -11,16 +11,16 @@
         public const string FunctionalTestProjectSuffix = "FunctionalTests";
         // public const string ApiProjectSuffix = "WebApi";
         public const string ApiProjectSuffix = "";
-        public const string MessagesProjName = "Messages";
+        public const string SharedKernelProjName = "SharedKernel";
 
         public static ClassPath SolutionClassPath(string solutionDirectory, string className)
         {
             return new ClassPath(solutionDirectory, "", className);
         }
 
-        public static ClassPath BaseMessageClassPath(string solutionDirectory, string className)
+        public static ClassPath MessagesClassPath(string solutionDirectory, string className)
         {
-            return new ClassPath(solutionDirectory, MessagesProjName, className);
+            return new ClassPath(solutionDirectory, Path.Combine(SharedKernelProjName, "Messages"), className);
         }
 
         public static ClassPath ControllerClassPath(string solutionDirectory, string className, string projectBaseName, string version = "v1")
@@ -32,6 +32,11 @@
         public static ClassPath UnitTestClassPath(string solutionDirectory, string className, string projectBaseName)
         {
             return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}.{UnitTestProjectSuffix}", "UnitTests"), className);
+        }
+
+        public static ClassPath UnitTestEntityTestsClassPath(string testDirectory, string className, string entityPlural, string projectBaseName)
+        {
+            return new ClassPath(testDirectory, Path.Combine($"{projectBaseName}.{UnitTestProjectSuffix}", "UnitTests", "Domain", entityPlural), className);
         }
 
         public static ClassPath UnitTestWrapperTestsClassPath(string solutionDirectory, string className, string projectBaseName)
@@ -185,6 +190,11 @@
             return new ClassPath(projectDirectory, Path.Combine($"{authServerProjectName}"), className);
         }
 
+        public static ClassPath ServicesTestClassPath(string testDirectory, string className, string projectBaseName)
+        {
+            return new ClassPath(testDirectory, Path.Combine($"{projectBaseName}.{IntegrationTestProjectSuffix}", "ServiceTests"), className);
+        }
+
         public static ClassPath FeatureTestClassPath(string solutionDirectory, string className, string entityName, string projectBaseName)
         {
             return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}.{IntegrationTestProjectSuffix}", "FeatureTests", entityName), className);
@@ -218,10 +228,10 @@
             return new ClassPath(srcDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Databases"), className);
         }
 
-        public static ClassPath ValidationClassPath(string solutionDirectory, string className, string entityPlural, string projectBaseName)
+        public static ClassPath ValidationClassPath(string srcDirectory, string className, string entityPlural, string projectBaseName)
         {
             var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
-            return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Domain", entityPlural, "Validators"), className);
+            return new ClassPath(srcDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Domain", entityPlural, "Validators"), className);
         }
 
         public static ClassPath ProfileClassPath(string solutionDirectory, string className, string entityPlural, string projectBaseName)
@@ -230,37 +240,33 @@
             return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Domain", entityPlural, "Mappings"), className);
         }
 
-        public static ClassPath FeaturesClassPath(string solutionDirectory, string className, string entityName, string projectBaseName)
+        public static ClassPath FeaturesClassPath(string srcDirectory, string className, string entityName, string projectBaseName)
         {
             var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
             return string.IsNullOrEmpty(entityName)
-                ? new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Domain"), className)
-                : new ClassPath(solutionDirectory,
+                ? new ClassPath(srcDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Domain"), className)
+                : new ClassPath(srcDirectory,
                     Path.Combine($"{projectBaseName}{withSuffix}", "Domain", entityName, "Features"), className);
         }
 
-        public static ClassPath ConsumerFeaturesClassPath(string solutionDirectory, string className, string projectBaseName)
+        public static ClassPath ConsumerFeaturesClassPath(string srcDirectory, string className, string domainDirectory, string projectBaseName)
         {
-            var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
-            return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Domain", "EventHandlers"), className);
+            return FeaturesClassPath(srcDirectory, className, domainDirectory, projectBaseName);
         }
 
-        public static ClassPath ProducerFeaturesClassPath(string solutionDirectory, string className, string projectBaseName)
+        public static ClassPath ProducerFeaturesClassPath(string srcDirectory, string className, string domainDirectory, string projectBaseName)
         {
-            var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
-            return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Domain", "EventHandlers"), className);
+            return FeaturesClassPath(srcDirectory, className, domainDirectory, projectBaseName);
         }
 
-        public static ClassPath ExceptionsClassPath(string solutionDirectory, string className, string projectBaseName)
+        public static ClassPath ExceptionsClassPath(string solutionDirectory, string className)
         {
-            var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
-            return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Exceptions"), className);
+            return new ClassPath(solutionDirectory, Path.Combine(SharedKernelProjName, "Exceptions"), className);
         }
 
-        public static ClassPath WrappersClassPath(string solutionDirectory, string className, string projectBaseName)
+        public static ClassPath WrappersClassPath(string srcDirectory, string className, string projectBaseName)
         {
-            var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
-            return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Wrappers"), className);
+            return new ClassPath(srcDirectory, Path.Combine($"{projectBaseName}", "Wrappers"), className);
         }
 
         public static ClassPath WebApiResourcesClassPath(string srcDirectory, string className, string projectBaseName)
@@ -271,6 +277,18 @@
         public static ClassPath WebApiServicesClassPath(string srcDirectory, string className, string projectBaseName)
         {
             return new ClassPath(srcDirectory, Path.Combine($"{projectBaseName}", "Services"), className);
+        }
+
+        public static ClassPath PolicyDomainClassPath(string srcDirectory, string className, string projectBaseName)
+        {
+            // wanted to have a `Policy` directory under Domain for permissions, roles, and rolepermissions, but changing the class path for
+            // rolepermissions scaffolding wasn't worth the lift atm. so this is currently the same as domain, but leaving it in hopes of an easier refactor in the future.
+            return new ClassPath(srcDirectory, Path.Combine($"{projectBaseName}", "Domain"), className);
+        }
+
+        public static ClassPath SharedKernelDomainClassPath(string solutionDirectory, string className)
+        {
+            return new ClassPath(solutionDirectory, Path.Combine($"{SharedKernelProjName}", "Domain"), className);
         }
 
         public static ClassPath WebApiProjectRootClassPath(string solutionDirectory, string className, string projectBaseName)
@@ -289,9 +307,9 @@
             return new ClassPath(solutionDirectory, $"{projectBaseName}.{UnitTestProjectSuffix}", className);
         }
 
-        public static ClassPath MessagesProjectRootClassPath(string solutionDirectory, string className)
+        public static ClassPath SharedKernelProjectRootClassPath(string solutionDirectory, string className)
         {
-            return new ClassPath(solutionDirectory, MessagesProjName, className);
+            return new ClassPath(solutionDirectory, SharedKernelProjName, className);
         }
         
         public static ClassPath ExampleYamlRootClassPath(string solutionDirectory, string className)
@@ -311,14 +329,13 @@
 
         public static ClassPath DtoClassPath(string solutionDirectory, string className, string entityName, string projectBaseName)
         {
-            var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
-            return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Dtos", entityName), className);
+            return new ClassPath(solutionDirectory, Path.Combine(SharedKernelProjName, "Dtos", projectBaseName, entityName), className);
         }
 
-        public static ClassPath SharedDtoClassPath(string solutionDirectory, string className, string projectBaseName)
+        public static ClassPath SharedDtoClassPath(string solutionDirectory, string className)
         {
             var withSuffix = ApiProjectSuffix.Length > 0 ? $".{ApiProjectSuffix}" : "";
-            return new ClassPath(solutionDirectory, Path.Combine($"{projectBaseName}{withSuffix}", "Dtos", "Shared"), className);
+            return new ClassPath(solutionDirectory, Path.Combine(SharedKernelProjName, "Dtos", "Shared"), className);
         }
 
         public static ClassPath IntegrationTestProjectClassPath(string solutionDirectory, string projectBaseName)
@@ -341,9 +358,9 @@
             return new ClassPath(solutionDirectory, $"{projectBaseName}.{UnitTestProjectSuffix}", $"{projectBaseName}.{UnitTestProjectSuffix}.csproj");
         }
 
-        public static ClassPath MessagesProjectClassPath(string solutionDirectory)
+        public static ClassPath SharedKernelProjectClassPath(string solutionDirectory)
         {
-            return new ClassPath(solutionDirectory, MessagesProjName, $"{MessagesProjName}.csproj");
+            return new ClassPath(solutionDirectory, SharedKernelProjName, $"{SharedKernelProjName}.csproj");
         }
 
         public static ClassPath WebApiProjectClassPath(string projectDirectory, string projectBaseName)
