@@ -49,8 +49,13 @@
             get => _responseType;
             set => _responseType = Utilities.PropTypeCleanup(value);
         }
-        
-        public List<Policy> Policies { get; set; } = new List<Policy>();
+
+        private string _permission;
+        public string PermissionName
+        {
+            get => _permission ?? $"Can{Name}";
+            set => _permission = value;
+        }
 
         private string _batchPropName;
         /// <summary>
@@ -86,6 +91,11 @@
         /// The name of the DbSet for the FK you are doing a batch add on. Generally, the plural of the FK entity. Leave null if you're not batching on a FK.
         /// </summary>
         public string BatchPropertyDbSetName { get; set; }
+
+        /// <summary>
+        /// Determines whether or not the feature is protected with an authorization policy attribute.
+        /// </summary>
+        public bool IsProtected { get; set; } = false;
 
         // feature role as command, producer, consumer in the future... dropped the ball on the OG implementation
     }
